% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrending-quarter.R, R/detrending.R
\name{adj_season_X13}
\alias{adj_season_X13}
\title{Removes the seasonality of a vector of quarterly values}
\usage{
adj_season_X13(x, start_date, remove_trend = FALSE)

adj_season_X13(x, start_date, remove_trend = FALSE)
}
\arguments{
\item{x}{vector of numerical values}

\item{start_date}{start date of the values}

\item{remove_trend}{should the estimated trend be removed? Default to ``FALSE}
}
\description{
Removes the seasonality of a vector of quarterly values

Removes the seasonality of a vector of monthly values
}
