% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_peru}
\alias{data_peru}
\title{Monthly Agriculture in Peru (2001-01 to 2015-12)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21960 rows and 210 columns.
}
\source{
<https://www.midagri.gob.pe/portal/boletin-estadistico-mensual-el-agro-en-cifras>
}
\usage{
data_peru
}
\description{
Monthly Agriculture in Peru (2001-01 to 2015-12)
}
\keyword{datasets}
