% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrending-quarter.R, R/detrending.R
\name{detrend_production}
\alias{detrend_production}
\title{Detrends the production data, using OLS}
\usage{
detrend_production(df, crop_name, region_id, in_log = FALSE)

detrend_production(df, crop_name, region_id, in_log = FALSE)
}
\arguments{
\item{df}{data}

\item{crop_name}{name of the crop}

\item{region_id}{id of the region}

\item{in_log}{if TRUE, the detrended values are expressed in log}
}
\value{
data frame with the product, the region id, the date, and the
  detrended value for the production. If `in_log = TRUE` and there are zeros
  the function returns `NULL`.

data frame with the product, the region id, the date, and the
  detrended value for the production. If `in_log = TRUE` and there are zeros
  the function returns `NULL`. Similarly, if there are more than two
}
\description{
Detrends the production data, using OLS

Detrends the production data, using OLS
}
