% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_minagri.R
\name{download.data}
\alias{download.data}
\title{Directly download the Excels versions of the monthly reports available on the
Peruvian website MINAGRI.}
\usage{
download.data(y)
}
\arguments{
\item{y}{last digits of the year (20yy)}

\item{dest_folder}{destination folder}
}
\description{
Directly download the Excels versions of the monthly reports available on the
Peruvian website MINAGRI.
}
