% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimations-quarter.R, R/estimations.R
\name{estimate_linear_lp}
\alias{estimate_linear_lp}
\title{Estimate Local Projections}
\usage{
estimate_linear_lp(
  df,
  horizons,
  y_name,
  group_name,
  crop_name,
  control_names,
  weather_names,
  add_month_fe = TRUE,
  add_intercept = FALSE,
  share_geo = NULL,
  transition_name = NULL,
  transition_method = NULL,
  state_names = c("planted", "harvested"),
  gamma = 3,
  std = c("nw", "cluster", "standard"),
  other_var_to_keep = NULL
)

estimate_linear_lp(
  df,
  horizons,
  y_name,
  group_name,
  crop_name,
  control_names,
  weather_names,
  add_month_fe = TRUE,
  add_intercept = FALSE,
  share_geo = NULL,
  transition_name = NULL,
  transition_method = NULL,
  state_names = c("planted", "harvested"),
  gamma = 3,
  std = c("nw", "cluster", "standard"),
  other_var_to_keep = NULL
)
}
\arguments{
\item{df}{original dataset}

\item{horizons}{number of horizons}

\item{y_name}{name of the exogenous variable}

\item{group_name}{name of the group variable}

\item{crop_name}{name of the crop to focus on}

\item{control_names}{vector of names of the control variables}

\item{weather_names}{vector of names of the weather variables}

\item{add_month_fe}{should columns with month dummy variables be added?
Default to `TRUE`}

\item{add_intercept}{should an intercept we added to the regressions?
(default to `FALSE`)}

\item{share_geo}{vector of names of the variables that contain the share of
each type of geographical pattern. By default `NULL`: no share used}

\item{transition_name}{name of the variable used to define the transition to
the two states. By default `NULL`}

\item{transition_method}{if transition function, name of the method to use:
`logistic` or `normal` (default to `NULL`, i.e., no transition)}

\item{state_names}{name of the two states in a vector of characters (only if
`transition_name` is not `NULL`). First period corresponds to mapped values
of `transition_name` close to 0, which is for large positive values of
`transition_name`}

\item{gamma}{logistic growth rate (default to 3, only used if
`transition_name` is not `NULL`)}

\item{std}{type of standard error (`"NW"` for Newey-West, `"Cluster"`,
`"Standard"` otherwise)}

\item{other_var_to_keep}{vector of names of other variables to keep in the
returned dataset (default to `NULL`: no additional vairable kept)}

\item{add_quarter_fe}{should columns with quarter dummy variables be added?
Default to `TRUE`}
}
\description{
Estimate Local Projections

Estimate Local Projections
}
