% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_minagri.R
\name{extract_pdf_data}
\alias{extract_pdf_data}
\title{Extract data from the PDF monthly reports from MINAGRI}
\usage{
extract_pdf_data(annee, mois, adresse, page, Cell1)
}
\arguments{
\item{annee}{year of the report}

\item{mois}{month of the report}

\item{adresse}{url of the report on www.midagri.gob.pe}

\item{page}{page number to extract}

\item{Cell1}{name of the first cell to import}
}
\description{
Extract data from the PDF monthly reports from MINAGRI
}
