% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_minagri.R
\name{format_header}
\alias{format_header}
\title{Format header (removing unecessary values)}
\usage{
format_header(x)
}
\arguments{
\item{x}{(vector) of string}
}
\description{
Format header (removing unecessary values)
}
