% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_index_longest_non_na}
\alias{get_index_longest_non_na}
\title{Returns the index of the longest sequence of non NA values in a vector}
\usage{
get_index_longest_non_na(y)
}
\arguments{
\item{y}{vector of numerical values}
}
\description{
Returns the index of the longest sequence of non NA values in a vector
}
