% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrending-quarter.R, R/detrending.R
\name{get_pct_dev_trend}
\alias{get_pct_dev_trend}
\title{Computes the percentage deviation from trend}
\usage{
get_pct_dev_trend(x, start_date)

get_pct_dev_trend(x, start_date)
}
\arguments{
\item{x}{vector of numerical monthly values}

\item{start_date}{start date of the values}
}
\description{
After estimating the trend and seasonal components of a quarterly
  time series using X-13 ARIMA-SEATS method, the percentage deviation of the
  series without seasonality to the trend id computed and returned in a
  tibble.

After estimating the trend and seasonal components of a monthly
  time series using X-13 ARIMA-SEATS method, the percentage deviation of the
  series without seasonality to the trend id computed and returned in a
  tibble.
}
