% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrending-quarter.R, R/detrending.R
\name{hp_filter_trend}
\alias{hp_filter_trend}
\title{Applies HP filter for quarterly data and returns the trend component}
\usage{
hp_filter_trend(x, freq = 14400)

hp_filter_trend(x, freq = 14400)
}
\arguments{
\item{x}{vector of monthly observations}
}
\description{
Applies HP filter for quarterly data and returns the trend component

Applies HP filter for monthly data and returns the trend component
}
