% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_minagri.R
\name{import_monthly_regional_values_SR}
\alias{import_monthly_regional_values_SR}
\title{Import monthly regional agricultural values from the Excel file,
for all the months of the file.}
\usage{
import_monthly_regional_values_SR(file, anneesup, timescale)
}
\arguments{
\item{file}{path to the Excel File}

\item{anneesup}{integer, if different from 1, keeping only the last year available}

\item{timescale}{}
}
\description{
Import monthly regional agricultural values from the Excel file,
for all the months of the file.
}
