% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_minagri.R
\name{import_monthly_regional_values_year_Px}
\alias{import_monthly_regional_values_year_Px}
\title{Import monthly regional agricultural values from the Excel file,
for a given month}
\usage{
import_monthly_regional_values_year_Px(sheet_name, file, anneesup)
}
\arguments{
\item{sheet_name}{name of the Excel sheet}

\item{file}{path to the Excel File}

\item{anneesup}{integer, if different from 1, keeping only the last year available}
}
\description{
Import monthly regional agricultural values from the Excel file,
for a given month
}
