% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_values_longest_non_na}
\alias{keep_values_longest_non_na}
\title{Returns a logical vector that identifies the longest sequence of non NA
values within the input vector}
\usage{
keep_values_longest_non_na(y)
}
\arguments{
\item{y}{numeric vector}
}
\description{
Returns a logical vector that identifies the longest sequence of non NA
values within the input vector
}
