% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_minagri.R
\name{missing_table}
\alias{missing_table}
\title{Importing data where table is missing for a given month by taking its value
in the excel file of the next year.}
\usage{
missing_table(y, mm, type = c("Production", "Superficies_R", "Superficies"))
}
\arguments{
\item{y}{last digit of the year (20..)}

\item{mm}{month (numeric)}

\item{type}{type of variable ("Production" for production, "Superficies_R"
for harvested surface, "Superficies" for planted surface)}
}
\description{
Importing data where table is missing for a given month by taking its value
in the excel file of the next year.
}
