% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_map_paper}
\alias{theme_map_paper}
\title{Theme for maps with ggplot2}
\usage{
theme_map_paper(...)
}
\arguments{
\item{...}{arguments passed to the theme function}
}
\description{
Theme for maps with ggplot2
}
