% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_map_paper_dark}
\alias{theme_map_paper_dark}
\title{Theme for maps with ggplot2 with a dark theme}
\usage{
theme_map_paper_dark(...)
}
\arguments{
\item{...}{arguments passed to the theme function}
}
\description{
Theme for maps with ggplot2 with a dark theme
}
