% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimations.R
\name{estimate_linear_lp}
\alias{estimate_linear_lp}
\title{Estimate Local Projections}
\usage{
estimate_linear_lp(
  df,
  horizons,
  y_name,
  group_name,
  crop_name,
  control_names,
  weather_names,
  share_geo = NULL,
  transition_name = NULL,
  gamma = 3,
  std = c("nw", "cluster", "standard"),
  other_var_to_keep = NULL
)
}
\arguments{
\item{df}{original dataset}

\item{horizons}{number of horizons}

\item{y_name}{name of the exogenous variable}

\item{group_name}{name of the group variable}

\item{crop_name}{name of the crop to focus on}

\item{control_names}{vector of names of the control variables}

\item{weather_names}{vector of names of the weather variables}

\item{share_geo}{vector of names of the variables that contain the share of
each type of geographical pattern. By default `NULL`: no share used}

\item{transition_name}{name of the variable used to define the transition to
the two states. By default `NULL`}

\item{gamma}{logistic growth rate (default to 3, only used if
`transition_name` is not `NULL`)}

\item{std}{type of standard error (`"NW"` for Newey-West, `"Cluster"`,
`"Standard"` otherwise)}

\item{other_var_to_keep}{vector of names of other variables to keep in the
returned dataset (default to `NULL`: no additional vairable kept)}
}
\description{
Estimate Local Projections
}
